extract atoms from editorviewr
